# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Customer Support Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"

  # ─── ORGANIZATIONS ──────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in the organization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: inviteUser
      tags: [users]
      summary: Invite a user to the organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInviteInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: number
        - name: tagId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, pending, snoozed]
        - name: channelId
          in: query
          schema:
            type: number
        - name: contactId
          in: query
          schema:
            type: number
        - name: assigneeId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
...
  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── CHANNELS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName, organizationName]
      properties:
        email:
          type: string

        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        organizationName:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResult:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── ORGANIZATION ─────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, agent]
        isActive:
          type: boolean
        organizationId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    UserInviteInput:
      type: object
      required: [email, firstName, lastName, role]
      properties:
        email:
          type: string

        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, agent]
...
    Contact:
      type: object
      required: [id, firstName, lastName, createdAt]
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        companyId:
          type: ["number", "null"]
        companyName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        lastActivityAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ContactInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: number
        tagIds:
          type: array
          items:
            type: number
...
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        domain:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        contactCount:
          type: number
        createdAt:
          type: string
          format: date-time
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        size:
          type: string
...
    Conversation:
      type: object
      required: [id, status, channelType, createdAt]
      properties:
        id:
          type: number
        status:
          type: string
          enum: [open, resolved, pending, snoozed]
        channelType:
          type: string
          enum: [whatsapp, messenger, instagram, telegram, tiktok, web]
        channelId:
          type: ["number", "null"]
        contactId:
          type: ["number", "null"]
        contact:
          $ref: "#/components/schemas/Contact"
        assigneeId:
          type: ["number", "null"]
        assignee:
          $ref: "#/components/schemas/User"
        subject:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        lastMessage:
          type: ["string", "null"]
        unreadCount:
          type: number
        aiHandled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [channelType]
      properties:
        channelType:
          type: string
          enum: [whatsapp, messenger, instagram, telegram, tiktok, web]
        channelId:
          type: number
        contactId:
          type: number
        subject:
          type: string
        assigneeId:
          type: number
...
    Message:
      type: object
      required: [id, conversationId, content, senderType, createdAt]
      properties:
        id:
          type: number
        conversationId:
          type: number
        content:
          type: string
        senderType:
          type: string
          enum: [contact, agent, ai, system]
        senderId:
          type: ["number", "null"]
        senderName:
          type: ["string", "null"]
        messageType:
          type: string
          enum: [text, image, file, audio, video]
        status:
          type: string
          enum: [sent, delivered, read, failed]
        isPrivate:
          type: boolean
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        messageType:
          type: string
          enum: [text, image, file, audio, video]
          default: text
        isPrivate:
          type: boolean

    # ─── CHANNEL ──────────────────────────────────────────────────────────────
...
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: number
        name:
          type: string
        color:
          type: string
...